Update to cstate v4
This updates the theme to cstate v4 and adds several config values that
are now required. In addition, it uses new features in v4 to update the
most recent planned maintenance issue to not affect the current status
of services so that if an issue occurs in the mean time it shows up and
doesn't require removing or "resolving" the planned maintenance.